RLP-933 Refactoring decorators and improving calls with a meta-decorator
